Organized a few things
